+++ packages/conf/src/types.ts
@@ -111,7 +111,7 @@ type LocaleObject = {
 
 export type FallbackLocales = LocaleObject
 
-type ModuleSource = [string, string?]
+type ModuleSource = readonly [module: string, specifier?: string]
 
 type CatalogService = {
   name: string
@@ -207,7 +207,9 @@ export type LinguiConfig = {
   orderBy?: OrderBy
   pseudoLocale?: string
   rootDir?: string
-  runtimeConfigModule?: ModuleSource | { [symbolName: string]: ModuleSource }
+  runtimeConfigModule?:
+    | ModuleSource
+    | Partial<Record<"useLingui" | "Trans" | "i18n", ModuleSource>>
   sourceLocale?: string
   service?: CatalogService
   experimental?: {
@@ -215,15 +217,16 @@ export type LinguiConfig = {
   }
 }
 
+type ModuleSourceNormalized = readonly [module: string, specifier: string]
+
 export type LinguiConfigNormalized = Omit<
   LinguiConfig,
   "runtimeConfigModule"
 > & {
   fallbackLocales?: FallbackLocales
   runtimeConfigModule: {
-    i18nImportModule: string
-    i18nImportName: string
-    TransImportModule: string
-    TransImportName: string
+    i18n: ModuleSourceNormalized
+    useLingui: ModuleSourceNormalized
+    Trans: ModuleSourceNormalized
   }
 }